optimization and corrections

+++ module_2/lesson_4/src/solution_1.py
@@ -2,4 +2,4 @@ import math
 #Запрашиваем float значение
 money = float(input("Введите значение:"))
 #Округляем и выводим
-print(round(money))
+print(float(round(money)))
